update sql
change classroom_table

+++ sql/eas_navicat.sql
@@ -77,7 +77,7 @@ DROP TABLE IF EXISTS `classroom_info`;
 CREATE TABLE `classroom_info` (
   `classroom_id` int(11) NOT NULL AUTO_INCREMENT,
   `classroom_addr` varchar(64) DEFAULT NULL,
-  `classroom_avai` tinyint(1) DEFAULT NULL,
+  `classroom_avai` int DEFAULT NULL,
   `classroom_type` varchar(64) DEFAULT NULL,
   `classroom_capacity` int(11) DEFAULT NULL,
   `classroom_remark` text,
